feat: add reset module
