+++ cypress/fixtures/data.json
@@ -1,18 +1,3 @@
 {
-  "todos": [
-    {
-      "id": 1,
-      "title": "first"
-    },
-    {
-      "id": 2,
-      "title": "second"
-    }
-  ],
-  "people": [
-    {
-      "name": "A person 5155",
-      "id": 1
-    }
-  ]
+  "todos": []
 }
